requirements.txt
matplotlib
sympy
scipy
streamlit
openpyxl
fpdf2